// apps/web/src/api/generated/schema.ts
export interface components {
  schemas: {
    AuditResponse: {
      /** Format: int64 */
      id: number | string
      /** Format: date-time */
      occurredAt: string
      eventType: string
      outcome: string
      actorUserId: null | string
      actorEmail: null | string
      subjectId: null | string
      ipAddress: null | string
      details: null | string
    }
    ChangePasswordRequest: {
      currentPassword: string
      newPassword: string
    }
    CreateUserRequest: {
      email: string
      displayName: null | string
      temporaryPassword: string
      roles: null | string[]
    }
    LoginRequest: {
      email: string
      password: string
    }
    PagedResponseOfAuditResponse: {
      items: components['schemas']['AuditResponse'][]
      /** Format: int32 */
      page: number | string
      /** Format: int32 */
      pageSize: number | string
      /** Format: int32 */
      total: number | string
    }
    PagedResponseOfRoleResponse: {
      items: components['schemas']['RoleResponse'][]
      /** Format: int32 */
      page: number | string
      /** Format: int32 */
      pageSize: number | string
      /** Format: int32 */
      total: number | string
    }
    PagedResponseOfUserResponse: {
      items: components['schemas']['UserResponse'][]
      /** Format: int32 */
      page: number | string
      /** Format: int32 */
      pageSize: number | string
      /** Format: int32 */
      total: number | string
    }
    ResetPasswordRequest: {
      temporaryPassword: string
    }
    RoleResponse: {
      id: string
      name: string
      isProtected: boolean
      permissions: string[]
    }
    SaveRoleRequest: {
      name: string
      permissions: string[]
    }
    SetActiveRequest: {
      isActive: boolean
    }
    SetRolesRequest: {
      roles: string[]
    }
    UpdateProfileRequest: {
      displayName: null | string
    }
    UserResponse: {
      id: string
      email: string
      displayName: null | string
      isActive: boolean
      mustChangePassword: boolean
      /** Format: date-time */
      createdAt: string
      /** Format: date-time */
      lastLoginAt: null | string
      roles: string[]
      permissions: string[]
    }
  }
  responses: never
  parameters: never
  requestBodies: never
  headers: never
  pathItems: never
}